# docs/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: team@testseer.io
    name: TestSeer Team
  description: "TestSeer analyses Java service codebases to produce a structured knowledge\
    \ graph of classes, endpoints, and outbound HTTP calls. Use the Query API to ask\
    \ questions about your services — reachability, impact analysis, shared type consumers\
    \ — and always get freshness metadata alongside results."
  license:
    name: Proprietary
  title: TestSeer API
  version: 0.1.0
servers:
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /v1/notifications/index-events:
    get:
      operationId: indexEvents
      parameters:
      - in: query
        name: serviceId
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SseEmitter"
          description: OK
      summary: SSE stream of index-complete events
      tags:
      - Notifications
  /v1/notifications/index-events/poll:
    get:
      operationId: pollIndexEvents
      parameters:
      - in: query
        name: serviceId
        required: false
        schema:
          type: string
      - in: query
        name: after
        required: false
        schema:
          type: string
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListIndexCompleteEvent"
          description: OK
      summary: Long-poll fallback for index-complete events
      tags:
      - Notifications
...
  /webhook/github:
    post:
      description: "Accepts GitHub `push` and `pull_request` webhook events. Validates\
        \ the HMAC-SHA256 signature, extracts changed files (including OpenAPI JSON\
        \ via GitHub PR files API when needed), and queues analysis jobs to Kafka.\
        \ Returns 202 Accepted on success."
      operationId: handleWebhook
      parameters:
      - description: "GitHub event type (push, pull_request)"
        in: header
        name: X-GitHub-Event
        required: true
        schema:
          type: string
      - description: HMAC-SHA256 signature for payload verification
        in: header
        name: X-Hub-Signature-256
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: Webhook received but ignored (unsupported event or action)
        "202":
          description: Webhook accepted and queued
        "401":
          description: Invalid HMAC signature
      summary: Receive GitHub webhook
      tags:
      - Webhook
components:
  schemas:
...
    IndexCompleteEvent:
      type: object
      properties:
        commitSha:
          type: string
        eventType:
          type: string
        indexedAt:
          type: string
          format: date-time
        jobId:
          type: string
        orgId:
          type: string
        repo:
          type: string
        scope:
          type: string
        serviceId:
          type: string
...
    ResponseEnvelopeListIndexCompleteEvent:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/IndexCompleteEvent"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    SseEmitter:
      type: object
      properties:
        timeout:
          type: integer
          format: int64
...